+++ OptimoveSDK/Sources/Classes/PreferenceCenter/Models/Channel.swift
@@ -1,0 +1,5 @@
+public enum Channel: Int, Codable {
+    case mobilePush = 489
+    case webPush = 490
+    case sms = 493
+}
