+++ OptimoveSDK/Sources/Classes/PreferenceCenter/Models/PreferenceCenterConfig.swift
@@ -1,0 +1,5 @@
+struct PreferenceCenterConfig: Codable {
+    var region: Region
+    var tenantId: Int
+    var brandGroupId: String
+}
